# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/usuarios": {
      "post": {
        "operationId": "UsersController_crear",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Usuario creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioRespuestaDto"
                }
              }
            }
          },
          "400": {
            "description": "Campo inválido (dni/correo)"
          },
          "401": {
            "description": "Sin cookie de sesión válida"
          },
          "403": {
            "description": "Rol distinto de administrador/director"
          },
          "409": {
            "description": "DNI/código/correo duplicado"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Crea un Usuario para cualquiera de los cinco roles",
        "tags": [
          "usuarios"
        ]
      },
      "get": {
        "operationId": "UsersController_listar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Listado de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsuarioRespuestaDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Filtro rol/estado desconocido"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Lista usuarios, con filtro opcional por rol y estado",
        "tags": [
          "usuarios"
        ]
      }
    },
    "/usuarios/{id}": {
      "get": {
        "operationId": "UsersController_obtenerPorId",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioRespuestaDto"
                }
              }
            }
          },
          "400": {
            "description": "id malformado"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Consulta un Usuario por id",
        "tags": [
          "usuarios"
        ]
      },
      "patch": {
        "operationId": "UsersController_actualizar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Usuario actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioRespuestaDto"
                }
              }
            }
          },
          "400": {
            "description": "Campo inválido (dni/correo)"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "409": {
            "description": "DNI/código/correo duplicado"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Actualiza datos básicos de un Usuario (nunca el estado)",
        "tags": [
          "usuarios"
        ]
      }
    },
...
    "/anios-escolares": {
      "post": {
        "operationId": "AniosEscolaresController_crear",
        "parameters": [],
        "responses": {
          "201": {
            "description": "AnioEscolar creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnioEscolarRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin cookie de sesión válida"
          },
          "403": {
            "description": "Rol distinto de administrador/director"
          },
          "409": {
            "description": "Nombre duplicado"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Crea un AnioEscolar con activo=false",
        "tags": [
          "anios-escolares"
        ]
      },
      "get": {
        "operationId": "AniosEscolaresController_listar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Listado de años escolares",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnioEscolarRespuestaDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Filtro activo desconocido"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Lista años escolares, con filtro opcional por activo",
        "tags": [
          "anios-escolares"
        ]
      }
    },
    "/anios-escolares/{id}": {
      "get": {
        "operationId": "AniosEscolaresController_obtenerPorId",
        "parameters": [],
        "responses": {
          "200": {
            "description": "AnioEscolar encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnioEscolarRespuestaDto"
                }
              }
            }
          },
          "400": {
            "description": "id malformado"
          },
          "404": {
            "description": "AnioEscolar no encontrado"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Consulta un AnioEscolar por id",
        "tags": [
          "anios-escolares"
        ]
      },
      "patch": {
        "operationId": "AniosEscolaresController_actualizar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "AnioEscolar actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnioEscolarRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "AnioEscolar no encontrado"
          },
          "409": {
            "description": "Nombre duplicado"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Actualiza el nombre de un AnioEscolar (nunca activo)",
        "tags": [
          "anios-escolares"
        ]
      },
      "delete": {
        "operationId": "AniosEscolaresController_eliminar",
        "parameters": [],
        "responses": {
          "204": {
            "description": "AnioEscolar eliminado"
          },
          "404": {
            "description": "AnioEscolar no encontrado"
          },
          "409": {
            "description": "Existen Seccion/Aula/Matricula/Configuracion dependientes"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Elimina físicamente un AnioEscolar sin dependientes",
        "tags": [
          "anios-escolares"
        ]
      }
    },
...
    "/configuracion": {
      "get": {
        "operationId": "ConfiguracionController_obtener",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Configuración institucional",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin cookie de sesión válida"
          },
          "403": {
            "description": "Rol distinto de administrador/director"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Consulta la configuración institucional",
        "tags": [
          "configuracion"
        ]
      },
      "put": {
        "operationId": "ConfiguracionController_actualizar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Configuración actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionRespuestaDto"
                }
              }
            }
          },
          "400": {
            "description": "Campo inválido (color, zona horaria o dominio)"
          },
          "401": {
            "description": "Sin cookie de sesión válida"
          },
          "403": {
            "description": "Rol distinto de administrador/director"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Actualiza la configuración institucional (merge parcial, auditado)",
        "tags": [
          "configuracion"
        ]
      }
    },
    "/configuracion/comite": {
      "get": {
        "operationId": "ConfiguracionController_listarComite",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Integrantes del comité",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsuarioRespuestaDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Sin cookie de sesión válida"
          },
          "403": {
            "description": "Rol distinto de administrador/director"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Lista los usuarios con rol comite",
        "tags": [
          "configuracion"
        ]
      }
    }
  },
  "info": {
    "title": "SEEI API",
    "description": "Walking skeleton del sistema SEEI (#1)",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UsuarioRespuestaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID del usuario"
          },
          "nombres": {
            "type": "string",
            "description": "Nombres completos"
          },
          "dni": {
            "type": "string",
            "description": "DNI"
          },
          "codigo": {
            "type": "string",
            "description": "Código institucional único"
          },
          "correo": {
            "type": "string",
            "description": "Correo electrónico"
          },
          "rol": {
            "type": "string",
            "description": "Rol del usuario",
            "enum": [
              "estudiante",
              "docente",
              "comite",
              "administrador",
              "director"
            ]
          },
          "estado": {
            "type": "string",
            "description": "Estado del usuario",
            "enum": [
              "activo",
              "inactivo",
              "bloqueado"
            ]
          },
          "creado_en": {
            "type": "string",
            "description": "Fecha de creación (ISO 8601)"
          }
        },
        "required": [
          "id",
          "nombres",
          "dni",
          "codigo",
          "correo",
          "rol",
          "estado",
          "creado_en"
        ]
      },
...
      "AnioEscolarRespuestaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID del año escolar"
          },
          "nombre": {
            "type": "string",
            "description": "Nombre único del año escolar"
          },
          "activo": {
            "type": "boolean",
            "description": "Indica si es el año escolar activo"
          }
        },
        "required": [
          "id",
          "nombre",
          "activo"
        ]
      },
...
      "ConfiguracionRespuestaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID de la fila de configuración"
          },
          "nombre": {
            "type": "string",
            "description": "Nombre de la institución",
            "nullable": true
          },
          "director": {
            "type": "string",
            "description": "Nombre del director o directora",
            "nullable": true
          },
          "color_primario": {
            "type": "string",
            "description": "Color primario en formato hex",
            "nullable": true
          },
          "color_secundario": {
            "type": "string",
            "description": "Color secundario en formato hex",
            "nullable": true
          },
          "zona_horaria": {
            "type": "string",
            "description": "Zona horaria IANA",
            "nullable": true
          },
          "dominios_google": {
            "description": "Dominios Google Workspace permitidos",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "logo_presente": {
            "type": "boolean",
            "description": "Si hay un logo institucional persistido"
          },
          "logo_mime": {
            "type": "string",
            "description": "MIME type del logo persistido, si existe",
            "nullable": true
          }
        },
        "required": [
          "id",
          "dominios_google",
          "logo_presente"
        ]
      }
    }
  }
}